i forgor

+++ include/kernel/string.h
@@ -1,8 +1,8 @@
 #ifndef _KERNEL_STRING_H_
 #define _KERNEL_STRING_H_
 
-#include <kernel/types.h>
+#include <kernel/common.h>
 
-void *memset(void *s, int c, size_t n);
+void *memset(void *s, u8 c, size_t n);
 
 #endif // _KERNEL_STRING_H_
