+++ kernel/string.c
@@ -1,8 +1,9 @@
+#include <kernel/common.h>
 #include <kernel/string.h>
-#include <kernel/types.h>
 
-void *memset(void *s, int c, size_t n) {
-	uint8_t *p = (uint8_t*)s;
+void *memset(void *s, u8 c, size_t n)
+{
+	u8 *p = (u8 *)s;
 	for (size_t i = 0; i < n; ++i)
 		p[i] = c;
 	return s;
